requirements.txt
flask
gevent
python-dotenv
edge-tts
art
